+++ crates/client/src/pages/mod.rs
@@ -2,7 +2,6 @@ mod admin;
 mod connection_manager;
 mod discover;
 mod library;
-mod plugin_manager;
 mod progress_popup;
 mod search;
 mod settings;
@@ -14,7 +13,6 @@ pub use admin::*;
 pub use connection_manager::*;
 pub use discover::*;
 pub use library::*;
-pub use plugin_manager::*;
 pub use progress_popup::*;
 pub use search::*;
 pub use settings::*;
