--- crates/spyglass-plugin/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "spyglass-plugin"
-version = "0.1.1"
-authors = ["Andrew Huynh <andrew@spyglass.fyi>"]
-description = "A small client-side library for writing spyglass plugins"
-homepage = "https://github.com/spyglass-search/spyglass/tree/main/crates/spyglass-plugin"
-repository = "https://github.com/spyglass-search/spyglass/tree/main/crates/spyglass-plugin"
-readme = "README.md"
-keywords = ["spyglass", "webassembly", "wasm", "plugins"]
-edition = "2021"
-license = "MIT"
-
-[dependencies]
-serde = { workspace = true }
-serde_json = { workspace = true }
-ron = { workspace = true }
-url = "2.2"
-
-[lib]
-name = "spyglass_plugin"
-path = "src/lib.rs"
-crate-type = ["lib"]
