--- crates/spyglass-plugin/README.md
@@ -1,3 +1,0 @@
-## spyglass-plugin
-
-A small client side library for writing spyglass plugins.
