--- crates/spyglass-plugin/src/consts.rs
@@ -1,6 +1,0 @@
-pub mod env {
-    pub const HOST_OS: &str = "HOST_OS";
-    pub const HOST_HOME_DIR: &str = "HOME_DIR";
-    pub const BASE_CONFIG_DIR: &str = "BASE_CONFIG_DIR";
-    pub const BASE_DATA_DIR: &str = "BASE_DATA_DIR";
-}
