+++ crates/spyglass-rpc/src/lib.rs
@@ -92,9 +92,6 @@ pub trait Rpc {
     #[method(name = "toggle_pause")]
     async fn toggle_pause(&self, is_paused: bool) -> Result<(), Error>;
 
-    #[method(name = "toggle_plugin")]
-    async fn toggle_plugin(&self, name: String, enabled: bool) -> Result<(), Error>;
-
     #[method(name = "uninstall_lens")]
     async fn uninstall_lens(&self, name: String) -> Result<(), Error>;
