+++ crates/spyglass/src/api/mod.rs
@@ -186,10 +186,6 @@ impl RpcServer for SpyglassRpc {
         handler::toggle_pause(self.state.clone(), is_paused).await
     }
 
-    async fn toggle_plugin(&self, name: String, enabled: bool) -> Result<(), Error> {
-        handler::toggle_plugin(self.state.clone(), name, enabled).await
-    }
-
     async fn uninstall_lens(&self, name: String) -> Result<(), Error> {
         handler::uninstall_lens(self.state.clone(), &self.config, &name).await
     }
