+++ crates/spyglass/src/lib.rs
@@ -4,6 +4,5 @@ pub mod documents;
 pub mod filesystem;
 pub mod pipeline;
 pub mod platform;
-pub mod plugin;
 pub mod state;
 pub mod task;
