+++ crates/spyglass/src/plugin/exports.rs
@@ -10,10 +10,6 @@ use entities::sea_orm::ModelTrait;
 use entities::sea_orm::QueryFilter;
 use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
 use serde::{Deserialize, Serialize};
-use spyglass_plugin::Authentication;
-use spyglass_plugin::DocumentUpdate;
-use spyglass_plugin::HttpMethod;
-use spyglass_plugin::{DocumentResult, PluginEvent};
 use spyglass_searcher::{RetrievedDocument, WriteTrait};
 use std::path::Path;
 use std::str::FromStr;
